udev: Add partition symlinks
The current rule only creates a symlink for each device. Extend it to create
symlinks for partitions too.

To accomplish that, switch from ATTR{nsid} to ENV{ID_NSID} which is set by
60-persistent-storage.rules. Alternatively, this could use ATTRS{nsid} to match
parent attributes. With that change we can branch in the symlink creation rules
and add a -partX suffix for partitions.

-partX was chosen instead of pX, because the device names already contain
hyphens and pX looks weird on the os symlink (osp1). SCSI disk on Azure also
have -partX symlinks, so we stay consistent.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>