udev: install rules based on pkg-config variable
udev is actually hardcoded to look for rules under `/etc`, `/run`,
`/usr/local/lib`, and `/usr/lib`. Other arbitrary prefixes are not
checked. Rather than assume one of these though, let's use the `udevdir`
pkg-config variable that is set when installing udev. If this cannot be
found, it falls back to the previous behaviour of using the given
prefix.

This requires bumping the minimum CMake version to 3.4, but that is
nearly 8½ years old now. Even CentOS 7 has a much newer version than
this via EPEL.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>