build(doc): generate manpages only with -DGENERATE_MANPAGES=1
As pointed out by Noah in #15, Debian packaging lints issues in the
generated manpage due to formatting emitted by older versions of pandoc:
https://github.com/jgm/pandoc/issues/9020

While changes in recent versions of pandoc have improved the output,
let's just take the generated output from pandoc 3.2 as a starting point
and maintain it manually for the time being.  This reduces friction for
packaging and gives us flexibility if we wanted to do something in the
manpage that pandoc doesn't support.

Allow manpages to be generated at build-time with -DGENERATE_MANPAGES=1
which will make it easy to sync.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>