feat(udev): drop rules.in template in favor of complete rules file
Instead of generating the rules from a template with variables,
replace @AZURE_NVME_ID_INSTALL_DIR@ with a hardcoded /usr/sbin path.
To ensure the path is correct during a build, replace the string
as needed.

This allows the rules to be copied in the majority of cases while
still ensuring the correct rules file is produced during a build.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>