feat(udev): extend rules for compatibility in SCSI and older NVMe environments
There are multiple combinations of disks depending on instance vm size
and configuration:

- NVMe-only disks in v6+ families

- Mixed disk with SCSI OS & resource disks alongside NVMe local disks
  e.g. LSv2/3

- Mixed disk with SCSI resource disk alongside NVMe OS and data disks
  e.g. Ebdsv5 with NVMe controller configured

- SCSI-only disks with most instances prior to v6

For SCSI OS disk, export as /dev/disk/azure/os similar to NVMe OS disk.

For SCSI resource disk, export as /dev/disk/azure/resource as already
done in the existing udev rules implementations. Ideally we'd expose
SCSI local disks in the same manner as local NVMe disks, but since we
cannot be sure of what environment we're in, /dev/disk/azure/local
will be limited to local NVMe disks of any configuration.

For older NVMe local disk implementation such as LSv2/v3 which do not
expose identifiers for index, name, etc. export new links for
/dev/disk/azure/local/by-serial/<serial>.  Based on observation, it
is possible this serial contains the equivalent index identifier so
this may be revisited in the future if we determine it to be
guaranteed to be useful as such.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>