feat(azure-nvme-id): identify MSFT NVMe Accelerator v1.0 disks
These disks do not support vendor-specific id but is used for all v6
VM sizes.  For usability, show the type and lun in the same manner
as we do in the udev rules.

Note that this does not change the udev rules. azure-nvme-id will not
be invoked for this controller as there is no benefit to the cost of
performing IDENTIFY_NAMESPACE on these namespaces at this time.  This
may change in the future.

This leaves Microsoft NVMe Direct Disk (v1) without identification
information, but that's limited to older skus.  In this case there is
no alternative to provide a predictable name or index.  The udev
rules will continue to map them by id.

Fix some missing f-strings in selftest too.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>